feat: simple contains two compilation units
