+++ examples/Makefile
@@ -1,13 +1,17 @@
 CC = gcc
+LD = gcc
 CFLAGS = -Os -g3
 SRCS = $(wildcard *.c)
-TARGET = $(SRCS:%.c=%)
+OBJS = $(SRCS:%.c=%)
+TARGET = simple
 
 .PHONY: all
 all: $(TARGET)
+simple: $(OBJS)
+	$(LD) -o $@ $^
 
 %: %.c
-	$(CC) $(CFLAGS) -o $@ $<
+	$(CC) $(CFLAGS) -c -o $@ $<
 
 .PHONY: clean
 clean:
