--- examples/simple.c
@@ -1,12 +1,0 @@
-typedef struct hoge {
-  int hoge;
-  char hogehoge;
-  int array[2];
-} Hoge;
-
-Hoge hoges[3];
-
-int main(void) {
-  return 0;
-}
-
