+++ examples/simple1.c
@@ -1,0 +1,23 @@
+#include "simple2.h"
+
+typedef struct hoge {
+  int *hoge;
+  char hogehoge;
+  int array[2];
+  unsigned int fuga: 1;
+  struct student *doredore;
+} Hoge;
+
+union book {
+  char name[16];
+  int price;
+};
+
+union book bk;
+Hoge hoges[3];
+
+int main(void) {
+  sub1(c);
+  return 0;
+}
+
