+++ examples/simple2.c
@@ -1,0 +1,7 @@
+#include "simple2.h"
+
+int c = 3;
+
+int sub1(int i) {
+  return i;
+}
